feat: Add collapsed property to sidebar categories
This commit adds the collapsed property to the sidebar categories in order to allow collapsing and expanding of category sections. The collapsed property is used to control the initial state of the category, and it is set to true for the "Get started" category and the API category. This change enables a more compact and organized sidebar navigation experience for users.

Refs #279